+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.0 2021.06.11
+ - Moved cmd to jsonui-cmd
+ - Made package out of jsonui, enabling others with json files to use
+   Easy: just run jsonui.Interactive([]byte with json.)
 ## 1.0.1 2018.09.15
  - fixes `null` rendering
  - adds binary builds